# [aten] 8 bytes aligned vector loads for bf16 and fp16 dtypes in torch.cat