@ngimel 2025-04-03
Did you try actually launching half kernels with the same 128-bit loads? Is your improvement just due to vectorized path that narrow types didn't take previously, and you don't need extra 64-bit instantiation?