@zhaozhul 2025-04-03
> Did you try actually launching half kernels with the same 128-bit loads? Is your improvement just due to vectorized path that narrow types didn't take previously, and you don't need extra 64-bit instantiation?

Thanks Natalia. We originally tried to enable such vectorized load with 128-bit as well, but there was major perf regression:


for fp16 types, torch.cat is even slower than copy. This might be related to increased register access (since the byte-per-element is half of fp32, hence double the register access during vectorized loads)

```
torch-cat-D1-14318-D2-479-D3-843-dtype-torch.float16: 8 reg
         B  pt_eager      copy
0    100.0  0.027242  0.034411
1   1000.0  0.200789  0.144489
2  10000.0  1.934080  1.230432
3  20000.0  3.862880  2.433824

torch-cat-D1-46070-D2-493-D3-526-dtype-torch.float32: 4 reg
         B  pt_eager      copy
0    100.0  0.022901  0.035566
1   1000.0  0.133395  0.206304
2  10000.0  1.317136  1.843776
3  20000.0  2.708800  3.693952

torch-cat-D1-14655-D2-116-D3-770-dtype-torch.float64: 2 reg
         B  pt_eager      copy
0    100.0  0.033610  0.048977
1   1000.0  0.253016  0.355499
2  10000.0  2.527584  3.365216
3  20000.0  5.111072  6.726208
```