@ngimel 2025-04-03
Registers are 4 byte, and the load instruction should use fewer of them. But maybe index computations exert register pressure.