@zhaozhul 2025-04-03
> Registers are 4 byte, and the load instruction should use fewer of them. But maybe index computations exert register pressure.

Yes the kILP doubled due to smaller byte size here, hence more register pressure

```
    IndexType v_elementOffset[kILP];
    T reg_data[kILP];
```

https://github.com/pytorch/pytorch/pull/150233/commits/8ac179465ff6ca736162e9fa7c4f5ef96368372a#diff-c5d250b30de1c137c7be334acce924451e59a89f7135467204b30fa1b2462e0aL256